update properties ui
