+++ editor/components/ui-editor/action-group.tsx
@@ -1,0 +1,98 @@
+"use client";
+
+import * as React from "react";
+import { Slot } from "@radix-ui/react-slot";
+import { cva, type VariantProps } from "class-variance-authority";
+import { cn } from "@/components/lib/utils/index";
+
+const actionVariants = cva(
+  "inline-flex items-center justify-center gap-2 rounded-md text-sm font-medium hover:bg-muted hover:text-muted-foreground disabled:pointer-events-none disabled:opacity-50 data-[state=on]:bg-accent data-[state=on]:text-accent-foreground [&_svg]:pointer-events-none [&_svg:not([class*='size-'])]:size-4 [&_svg]:shrink-0 focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px] outline-none transition-[color,box-shadow] aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive whitespace-nowrap",
+  {
+    variants: {
+      variant: {
+        default: "bg-transparent",
+        outline:
+          "border border-input bg-transparent shadow-xs hover:bg-accent hover:text-accent-foreground",
+      },
+      size: {
+        default: "h-9 px-4 py-2 has-[>svg]:px-3",
+        xs: "h-6 rounded-md gap-1.5 px-2.5 has-[>svg]:px-2.5 text-xs",
+        sm: "h-8 rounded-md gap-1.5 px-3 has-[>svg]:px-2.5",
+        lg: "h-10 rounded-md px-6 has-[>svg]:px-4",
+        icon: "size-6 p-1 [&_svg:not([class*='size-'])]:size-3.5",
+      },
+    },
+    defaultVariants: {
+      variant: "default",
+      size: "default",
+    },
+  }
+);
+
+const ToggleGroupContext = React.createContext<
+  VariantProps<typeof actionVariants>
+>({
+  size: "default",
+  variant: "default",
+});
+
+function ActionGroup({
+  className,
+  variant,
+  size,
+  children,
+  ...props
+}: React.ComponentProps<"div"> & VariantProps<typeof actionVariants>) {
+  return (
+    <div
+      data-slot="action-group"
+      data-variant={variant}
+      data-size={size}
+      className={cn(
+        "group/action-group flex w-fit items-center rounded-md data-[variant=outline]:shadow-xs",
+        className
+      )}
+      {...props}
+    >
+      <ToggleGroupContext.Provider value={{ variant, size }}>
+        {children}
+      </ToggleGroupContext.Provider>
+    </div>
+  );
+}
+
+function ActionGroupItem({
+  className,
+  children,
+  variant,
+  size,
+  asChild,
+  ...props
+}: React.ComponentProps<"button"> &
+  VariantProps<typeof actionVariants> & {
+    asChild?: boolean;
+  }) {
+  const context = React.useContext(ToggleGroupContext);
+  const Comp = asChild ? Slot : "button";
+
+  return (
+    <Comp
+      data-slot="action-group-item"
+      data-variant={context.variant || variant}
+      data-size={context.size || size}
+      className={cn(
+        actionVariants({
+          variant: context.variant || variant,
+          size: context.size || size,
+        }),
+        "min-w-0 flex-1 shrink-0 rounded-none shadow-none first:rounded-l-md last:rounded-r-md focus:z-10 focus-visible:z-10 data-[variant=outline]:border-l-0 data-[variant=outline]:first:border-l",
+        className
+      )}
+      {...props}
+    >
+      {children}
+    </Comp>
+  );
+}
+
+export { ActionGroup, ActionGroupItem };
