+++ editor/scaffolds/sidecontrol/controls/ext-align.tsx
@@ -1,3 +1,7 @@
+import {
+  ActionGroup,
+  ActionGroupItem,
+} from "@/components/ui-editor/action-group";
 import { Button } from "@/components/ui-editor/button";
 import {
   Tooltip,
@@ -15,11 +19,13 @@ import {
   SpaceEvenlyVerticallyIcon,
 } from "@radix-ui/react-icons";
 import React from "react";
+import { cn } from "@/components/lib/utils/index";
 
 export function AlignControl({
   disabled,
   onAlign,
   onDistributeEvenly,
+  className,
 }: {
   disabled?: boolean;
   onAlign?: (alignment: {
@@ -27,68 +33,15 @@ export function AlignControl({
     vertical?: "none" | "min" | "max" | "center";
   }) => void;
   onDistributeEvenly?: (axis: "x" | "y") => void;
+  className?: string;
 }) {
   return (
-    <div className="w-full flex items-center justify-center gap-2">
-      <div className="flex items-center gap-0 justify-center">
-        <Item
-          disabled={disabled}
-          tooltip="Align left"
-          onClick={() => {
-            onAlign?.({ horizontal: "min" });
-          }}
-        >
-          <AlignLeftIcon />
-        </Item>
-        <Item
-          disabled={disabled}
-          tooltip="Align horizontal center"
-          onClick={() => {
-            onAlign?.({ horizontal: "center" });
-          }}
-        >
-          <AlignCenterHorizontallyIcon />
-        </Item>
-        <Item
-          disabled={disabled}
-          tooltip="Align right"
-          onClick={() => {
-            onAlign?.({ horizontal: "max" });
-          }}
-        >
-          <AlignRightIcon />
-        </Item>
-      </div>
-      <div className="flex items-center gap-0 justify-center">
-        <Item
-          disabled={disabled}
-          tooltip="Align top"
-          onClick={() => {
-            onAlign?.({ vertical: "min" });
-          }}
-        >
-          <AlignTopIcon />
-        </Item>
-        <Item
-          disabled={disabled}
-          tooltip="Align vertical center"
-          onClick={() => {
-            onAlign?.({ vertical: "center" });
-          }}
-        >
-          <AlignCenterVerticallyIcon />
-        </Item>
-        <Item
-          disabled={disabled}
-          tooltip="Align bottom"
-          onClick={() => {
-            onAlign?.({ vertical: "max" });
-          }}
-        >
-          <AlignBottomIcon />
-        </Item>
-      </div>
-      <div className="flex items-center gap-0 justify-center">
+    <div
+      className={cn("w-full flex items-center justify-center gap-1", className)}
+    >
+      <HorizontalItems disabled={disabled} onAlign={onAlign} />
+      <VerticalItems disabled={disabled} onAlign={onAlign} />
+      <ActionGroup size="icon" variant="outline">
         <Item
           disabled={disabled}
           tooltip="Distribute horizontally"
@@ -107,8 +60,92 @@ export function AlignControl({
         >
           <SpaceEvenlyVerticallyIcon />
         </Item>
-      </div>
+      </ActionGroup>
     </div>
+  );
+}
+
+function HorizontalItems({
+  disabled,
+  onAlign,
+}: {
+  disabled?: boolean;
+  onAlign?: (alignment: {
+    horizontal?: "none" | "min" | "max" | "center";
+  }) => void;
+}) {
+  return (
+    <ActionGroup size="icon" variant="outline">
+      <Item
+        disabled={disabled}
+        tooltip="Align left"
+        onClick={() => {
+          onAlign?.({ horizontal: "min" });
+        }}
+      >
+        <AlignLeftIcon />
+      </Item>
+      <Item
+        disabled={disabled}
+        tooltip="Align horizontal center"
+        onClick={() => {
+          onAlign?.({ horizontal: "center" });
+        }}
+      >
+        <AlignCenterHorizontallyIcon />
+      </Item>
+      <Item
+        disabled={disabled}
+        tooltip="Align right"
+        onClick={() => {
+          onAlign?.({ horizontal: "max" });
+        }}
+      >
+        <AlignRightIcon />
+      </Item>
+    </ActionGroup>
+  );
+}
+
+function VerticalItems({
+  disabled,
+  onAlign,
+}: {
+  disabled?: boolean;
+  onAlign?: (alignment: {
+    vertical?: "none" | "min" | "max" | "center";
+  }) => void;
+}) {
+  return (
+    <ActionGroup size="icon" variant="outline">
+      <Item
+        disabled={disabled}
+        tooltip="Align top"
+        onClick={() => {
+          onAlign?.({ vertical: "min" });
+        }}
+      >
+        <AlignTopIcon />
+      </Item>
+      <Item
+        disabled={disabled}
+        tooltip="Align vertical center"
+        onClick={() => {
+          onAlign?.({ vertical: "center" });
+        }}
+      >
+        <AlignCenterVerticallyIcon />
+      </Item>
+      <Item
+        disabled={disabled}
+        tooltip="Align bottom"
+        onClick={() => {
+          onAlign?.({ vertical: "max" });
+        }}
+      >
+        <AlignBottomIcon />
+      </Item>
+    </ActionGroup>
   );
 }
 
@@ -125,15 +162,13 @@ function Item({
   return (
     <Tooltip>
       <TooltipTrigger asChild>
-        <Button
-          variant="ghost"
-          size="icon"
-          className="p-1"
+        <ActionGroupItem
           disabled={disabled}
           onClick={onClick}
+          className="[&_svg:not([class*='size-'])]:size-3"
         >
           {children}
-        </Button>
+        </ActionGroupItem>
       </TooltipTrigger>
       <TooltipContent>{tooltip}</TooltipContent>
     </Tooltip>
