+++ editor/scaffolds/sidecontrol/controls/ext-mask.tsx
@@ -1,0 +1,10 @@
+import { Button } from "@/components/ui-editor/button";
+import { MaskOffIcon } from "@radix-ui/react-icons";
+
+export function MaskControl() {
+  return (
+    <Button variant="ghost" size="icon">
+      <MaskOffIcon />
+    </Button>
+  );
+}
